update product review view so superuser can delete reviews

+++ products/views.py
@@ -221,6 +221,13 @@ def delete_review(request, review_id):
     Delete a product review
     """
     review = get_object_or_404(Review, pk=review_id)
-    review.delete()
-    messages.success(request, 'Review deleted!')
-    return redirect(reverse('product_detail', args=(review.product.id,)))
+
+    if request.user == review.user or request.user.is_superuser:
+        review.delete()
+        messages.info(request, "Review deleted!")
+        return redirect(reverse("product_detail", args=[review.product.id]))
+    else:
+        messages.error(
+            request,
+            "Only store owner and the reviewer can do that.")
+        return redirect(reverse("product_detail", args=[review.product.id]))
